simple signed remaining gas

+++ crates/vm/levm/src/execution_handlers.rs
@@ -54,7 +54,7 @@ impl<'a> VM<'a> {
                     result: TxResult::Revert(error),
                     gas_used: callframe
                         .gas_limit
-                        .checked_sub(callframe.gas_remaining)
+                        .checked_sub(callframe.gas_remaining as u64)
                         .ok_or(InternalError::Underflow)?,
                     output: Bytes::new(),
                 });
@@ -72,7 +72,7 @@ impl<'a> VM<'a> {
                 let callframe = &mut self.current_call_frame;
                 callframe
                     .gas_limit
-                    .checked_sub(callframe.gas_remaining)
+                    .checked_sub(callframe.gas_remaining as u64)
                     .ok_or(InternalError::Underflow)?
             },
             output: std::mem::take(&mut self.current_call_frame.output),
@@ -96,7 +96,7 @@ impl<'a> VM<'a> {
             result: TxResult::Revert(error),
             gas_used: callframe
                 .gas_limit
-                .checked_sub(callframe.gas_remaining)
+                .checked_sub(callframe.gas_remaining as u64)
                 .ok_or(InternalError::Underflow)?,
             output: std::mem::take(&mut callframe.output),
         })
